# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Aethermind Agent OS API
  description: Multi-agent orchestration platform with LLM providers
  version: 0.1.0
  contact:
    name: Aethermind Support
    url: https://github.com/aethermind/agentos

servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.aethermind.io
    description: Production server

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        model:
          type: string
        status:
          type: string
          enum: [idle, running, paused, stopped]
        config:
          type: object

    Execution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        input:
          type: object
        output:
          type: object
        error:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer

    Log:
      type: object
      properties:
        id:
          type: string
          format: uuid
        executionId:
          type: string
          format: uuid
          nullable: true
        agentId:
          type: string
          format: uuid
          nullable: true
        level:
          type: string
          enum: [debug, info, warn, error]
        message:
          type: string
        metadata:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time

    Cost:
      type: object
      properties:
        id:
          type: string
          format: uuid
        executionId:
          type: string
          format: uuid
        model:
          type: string
        promptTokens:
          type: integer
        completionTokens:
          type: integer
        totalTokens:
          type: integer
        cost:
          type: number
          format: double
        currency:
          type: string
        createdAt:
          type: string
          format: date-time

    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        definition:
          type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        suggestion:
          type: string

    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean

paths:
  /health:
    get:
      summary: Health check
      security: []
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  storage:
                    type: string

  /api/agents:
    get:
      summary: List all agents
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - model
              properties:
                name:
                  type: string
                model:
                  type: string
                provider:
                  type: string
                config:
                  type: object
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/agents/{id}:
    get:
      summary: Get agent by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          description: Agent not found

    delete:
      summary: Delete agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Agent deleted
        '404':
          description: Agent not found

  /api/agents/{id}/execute:
    post:
      summary: Execute agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
      responses:
        '200':
          description: Execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'

  /api/logs:
    get:
      summary: Get logs
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: agentId
          in: query
          schema:
            type: string
        - name: executionId
          in: query
          schema:
            type: string
        - name: offset
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Paginated logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  /api/costs:
    get:
      summary: Get cost analytics
      parameters:
        - name: executionId
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Cost data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

  /api/workflows:
    get:
      summary: List workflows
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'

    post:
      summary: Create workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - definition
              properties:
                name:
                  type: string
                description:
                  type: string
                definition:
                  type: object
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'

  /api/workflows/{id}/execute:
    post:
      summary: Execute workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                input:
                  type: object
      responses:
        '200':
          description: Workflow execution started
          content:
            application/json:
              schema:
                type: object
                properties:
                  executionId:
                    type: string
                  status:
                    type: string